+++ src/mympd_api/jukebox.c
@@ -123,6 +123,12 @@ sds mympd_api_jukebox_list(struct t_partition_state *partition_state, struct t_s
     unsigned entities_found = 0;
     unsigned real_limit = offset + limit;
     struct t_list *expr_list = parse_search_expression_to_list(expression, SEARCH_TYPE_SONG);
+    if (expr_list == NULL) {
+        sdsclear(buffer);
+        buffer = jsonrpc_respond_message(buffer, cmd_id, request_id,
+            JSONRPC_FACILITY_DATABASE, JSONRPC_SEVERITY_ERROR, "Invalid search expression");
+        return buffer;
+    }
     buffer = jsonrpc_respond_start(buffer, cmd_id, request_id);
     buffer = sdscat(buffer, "\"data\":[");
     bool print_stickers = check_get_sticker(partition_state->mpd_state->feat.stickers, &tagcols->stickers);